chore: move usePlotVisibility to utils
