+++ src/lib/utils/usePlotVisibility.jsx
